Create vague
This code uses matplotlib.pyplot to create a plot and animate a faceless person. The person is represented by a circle for the head and rectangles for the body and legs. The update function is called for each frame of the animation, which moves the person to the right by updating the x-positions of the body parts. The animation is created using FuncAnimation and displayed using plt.show().